Restructure project for Azure Functions deployment
- Migrated from monolithic `function_app.py` to per-function folder structure
- Added `__init__.py` and `function.json` for each function in V2-compliant format
- Removed unused packages such as langchain and pywin32
- Cleaned up `requirements.txt` and ensured compatibility with Linux-based Azure Functions runtime
- Verified function detection and Event Grid integration via manual testing
