+++ AzureFunctions-Python-SpeakerDiarization/.funcignore
@@ -1,0 +1,1 @@
+
