+++ AzureFunctions-Python-SpeakerDiarization/TriggerTranscriptionJob/function.json
@@ -1,0 +1,11 @@
+{
+  "bindings": [
+    {
+      "type": "eventGridTrigger",
+      "name": "event",
+      "direction": "in"
+    }
+  ],
+  "scriptFile": "__init__.py",
+  "entryPoint": "trigger_transcription_job"
+}
