+++ AzureFunctions-Python-api/GetFeedback/__init__.py
@@ -1,0 +1,61 @@
+"""
+GetFeedback関数
+フィードバック一覧取得エンドポイント
+"""
+
+import azure.functions as func
+import logging
+import json
+import traceback
+import sys
+import os
+
+# パスを追加してutilsモジュールをインポート可能にする
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+from utils.db_utils import execute_query
+
+# FunctionAppインスタンスの生成（1回のみ）
+app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)
+
+# ロガーの設定
+logger = logging.getLogger(__name__)
+
+@app.function_name(name="GetFeedback")
+@app.route(route="feedback", methods=["GET", "OPTIONS"])
+def get_feedback_func(req: func.HttpRequest) -> func.HttpResponse:
+    """フィードバック一覧取得エンドポイント"""
+    try:
+        if req.method == "OPTIONS":
+            headers = {
+                "Access-Control-Allow-Origin": "*",
+                "Access-Control-Allow-Methods": "GET, OPTIONS",
+                "Access-Control-Allow-Headers": "Content-Type",
+            }
+            return func.HttpResponse(status_code=204, headers=headers)
+
+        query = """
+            SELECT f.feedback_id, f.meeting_id, f.user_id, f.feedback_text, 
+                   f.inserted_datetime, f.updated_datetime, u.user_name 
+            FROM dbo.Feedback f 
+            JOIN dbo.Users u ON f.user_id = u.user_id
+        """
+        feedback_list = execute_query(query)
+
+        headers = {"Content-Type": "application/json", "Access-Control-Allow-Origin": "*"}
+        return func.HttpResponse(
+            json.dumps(feedback_list, ensure_ascii=False),
+            mimetype="application/json",
+            status_code=200,
+            headers=headers
+        )
+
+    except Exception as e:
+        logger.error(f"Get feedback error: {str(e)}")
+        logger.error(f"Error details: {traceback.format_exc()}")
+        headers = {"Content-Type": "application/json", "Access-Control-Allow-Origin": "*"}
+        return func.HttpResponse(
+            json.dumps({"error": f"Internal server error: {str(e)}"}, ensure_ascii=False),
+            mimetype="application/json",
+            status_code=500,
+            headers=headers
+        )
