+++ AzureFunctions-Python-api/GetUserById/__init__.py
@@ -1,0 +1,76 @@
+"""
+GetUserById関数
+ユーザー情報取得エンドポイント
+"""
+
+import azure.functions as func
+import logging
+import json
+import traceback
+import sys
+import os
+
+# パスを追加してutilsモジュールをインポート可能にする
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+from utils.db_utils import execute_query
+
+# FunctionAppインスタンスの生成（1回のみ）
+app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)
+
+# ロガーの設定
+logger = logging.getLogger(__name__)
+
+@app.function_name(name="GetUserById")
+@app.route(route="users/{user_id}", methods=["GET", "OPTIONS"])
+def get_user_by_id_func(req: func.HttpRequest) -> func.HttpResponse:
+    """ユーザー情報取得エンドポイント"""
+    try:
+        if req.method == "OPTIONS":
+            headers = {
+                "Access-Control-Allow-Origin": "*",
+                "Access-Control-Allow-Methods": "GET, OPTIONS",
+                "Access-Control-Allow-Headers": "Content-Type",
+            }
+            return func.HttpResponse(status_code=204, headers=headers)
+
+        user_id = req.route_params.get('user_id')
+        if not user_id:
+            return func.HttpResponse(
+                json.dumps({"error": "User ID is required"}, ensure_ascii=False),
+                mimetype="application/json",
+                status_code=400
+            )
+
+        # データベースからユーザー情報を取得
+        query = """
+            SELECT user_id, user_name, email, is_manager, manager_id, is_active, account_status 
+            FROM dbo.Users 
+            WHERE user_id = ?
+        """
+        users = execute_query(query, (user_id,))
+
+        if not users:
+            return func.HttpResponse(
+                json.dumps({"error": "User not found"}, ensure_ascii=False),
+                mimetype="application/json",
+                status_code=404
+            )
+
+        headers = {"Content-Type": "application/json", "Access-Control-Allow-Origin": "*"}
+        return func.HttpResponse(
+            json.dumps(users[0], ensure_ascii=False),
+            mimetype="application/json",
+            status_code=200,
+            headers=headers
+        )
+
+    except Exception as e:
+        logger.error(f"Get user error: {str(e)}")
+        logger.error(f"Error details: {traceback.format_exc()}")
+        headers = {"Content-Type": "application/json", "Access-Control-Allow-Origin": "*"}
+        return func.HttpResponse(
+            json.dumps({"error": f"Internal server error: {str(e)}"}, ensure_ascii=False),
+            mimetype="application/json",
+            status_code=500,
+            headers=headers
+        )
