+++ AzureFunctions-Python-api/SearchBasicInfo/__init__.py
@@ -1,0 +1,56 @@
+"""
+SearchBasicInfo関数
+基本情報検索エンドポイント
+"""
+
+import azure.functions as func
+import logging
+import json
+import traceback
+import sys
+import os
+
+# パスを追加してutilsモジュールをインポート可能にする
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+from utils.db_utils import execute_query
+
+# FunctionAppインスタンスの生成（1回のみ）
+app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)
+
+# ロガーの設定
+logger = logging.getLogger(__name__)
+
+@app.function_name(name="SearchBasicInfo")
+@app.route(route="basicinfo/search", methods=["GET", "OPTIONS"])
+def search_basic_info_func(req: func.HttpRequest) -> func.HttpResponse:
+    """基本情報検索エンドポイント"""
+    try:
+        if req.method == "OPTIONS":
+            headers = {
+                "Access-Control-Allow-Origin": "*",
+                "Access-Control-Allow-Methods": "GET, OPTIONS",
+                "Access-Control-Allow-Headers": "Content-Type",
+            }
+            return func.HttpResponse(status_code=204, headers=headers)
+
+        query = "SELECT * FROM dbo.BasicInfo"
+        basic_info = execute_query(query)
+
+        headers = {"Content-Type": "application/json", "Access-Control-Allow-Origin": "*"}
+        return func.HttpResponse(
+            json.dumps(basic_info, ensure_ascii=False),
+            mimetype="application/json",
+            status_code=200,
+            headers=headers
+        )
+
+    except Exception as e:
+        logger.error(f"Search basic info error: {str(e)}")
+        logger.error(f"Error details: {traceback.format_exc()}")
+        headers = {"Content-Type": "application/json", "Access-Control-Allow-Origin": "*"}
+        return func.HttpResponse(
+            json.dumps({"error": f"Internal server error: {str(e)}"}, ensure_ascii=False),
+            mimetype="application/json",
+            status_code=500,
+            headers=headers
+        )
