--- AzureFunctions-Python-api/consolelog.md
@@ -1,22 +1,0 @@
-page.tsx:83 
- 
- GET http://localhost:7071/api/basicinfo/65 404 (Not Found)
-app-index.tsx:25 
- Error fetching meeting: SyntaxError: Failed to execute 'json' on 'Response': Unexpected end of JSON input
-    at fetchMeeting (page.tsx:84:35)
-feedback.ts:10 
- 
- GET http://localhost:7071/api/conversation/segments/65 net::ERR_ABORTED 404 (Not Found)
-app-index.tsx:25 
- APIエラー: HTTP 404
-feedback.ts:10 
- 
- GET http://localhost:7071/api/conversation/segments/65 404 (Not Found)
-app-index.tsx:25 
- APIエラー: HTTP 404
-page.tsx:83 
- 
- GET http://localhost:7071/api/basicinfo/65 404 (Not Found)
-app-index.tsx:25 
- Error fetching meeting: SyntaxError: Failed to execute 'json' on 'Response': Unexpected end of JSON input
-    at fetchMeeting (page.tsx:84:35)
